Remove default model creation from file manager
All new models are created from templates.
